requirements.txt
asn1crypto==1.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coincurve==20.0.0
cryptography==44.0.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
nostr==0.0.2
pycparser==2.22
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
secp256k1==0.14.0
shellingham==1.5.4
tlv8==0.10.0
tornado==6.4.2
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
